+++ md2star/preprocessing/__init__.py
@@ -16,8 +16,8 @@ Author
 [Warith HARCHAOUI](https://linkedin.com/in/warith-harchaoui/)
 """
 
-from .alt_text import DEFAULT_ALT_TEXT_MODEL, fill_empty_alt_text
-from .lint import DEFAULT_LINT_MODEL, is_ollama_installed
+from .alt_text import fill_empty_alt_text
+from .lint import lint_with_llm
 from .mermaid import render_mermaid_local
 from .pipeline import PHASES, isolate_images_for_pptx, preprocess_markdown
 
@@ -26,8 +26,6 @@ __all__ = [
     "isolate_images_for_pptx",
     "render_mermaid_local",
     "fill_empty_alt_text",
-    "is_ollama_installed",
-    "DEFAULT_LINT_MODEL",
-    "DEFAULT_ALT_TEXT_MODEL",
+    "lint_with_llm",
     "PHASES",
 ]